pyproject.toml
[project]
authors = [
    {name = "Nick DeRobertis", email = "derobertis.nick@gmail.com"},
]
requires-python = "<4.0,>=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "livekit-agents>=0.12.18",
    "livekit-plugins-deepgram>=0.7.1",
    "livekit-plugins-elevenlabs>=0.8.1",
    "livekit-plugins-openai>=0.12.0",
    "livekit-plugins-silero>=0.7.5",
]
name = "spanish-language-tutor"
version = "0.1.0"
description = "A Spanish language tutor that uses LiveKit to provide real-time feedback on pronunciation and grammar."
readme = "README.md"

[dependency-groups]
dev = [
    "pyright",
]
